refactor: change datatype from double to float in main.cpp and update related calculations refactor: update argument names for blocks and threads in main.hpp to reflect their multipliers

+++ examples/vector_add_overlap/main.cpp
@@ -5,7 +5,7 @@
 
 #include "main.hpp"
 
-using datatype = double;
+using datatype = int;
 
 template <typename VT>
 void checkdata(const gcxx::span<VT>& h_a, VT checkval) {
